EventIndex: Move the event indexing files into a separate folder.
