+++ src/Searching.js
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import EventIndexPeg from "./EventIndexPeg";
+import EventIndexPeg from "./indexing/EventIndexPeg";
 import MatrixClientPeg from "./MatrixClientPeg";
 
 function serverSideSearch(term, roomId = undefined) {
