+++ src/indexing/BaseEventIndexManager.js
