+++ src/indexing/EventIndex.js
@@ -14,8 +14,8 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import PlatformPeg from "./PlatformPeg";
-import MatrixClientPeg from "./MatrixClientPeg";
+import PlatformPeg from "../PlatformPeg";
+import MatrixClientPeg from "../MatrixClientPeg";
 
 /**
  * Event indexing class that wraps the platform specific event indexing.
